fix (Chapter03) some errors has been corrected.

+++ src/Capitulo03/Bloque03/Ejercicio01_SumayMediaDeTodoslosNumeros.java
@@ -8,11 +8,11 @@ public class Ejercicio01_SumayMediaDeTodoslosNumeros {
 
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in); 
-		int num = 1, numSuma = 0, i; //declaramos la variante i ahora para utilizarla para realizar la media
+		float num = 1, numSuma = 0, i; //declaramos la variante i ahora para utilizarla para realizar la media
 		 
 		for (i = -1 ;num != 0; i++) {	//el valor de i será de -1 para que al introducir el número 0 se realize la media correctamente.
 			System.out.println("Introduce un número(0 para terminar)");
-			num = sc.nextInt();
+			num = sc.nextFloat();
 			
 			numSuma = num + numSuma;
 		}
